backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepguard-backend"
version = "0.1.0"
description = "FastAPI backend for DeepGuard"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "alembic>=1.13.0,<2.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-jose[cryptography]>=3.3.0,<4.0.0",
  "passlib[bcrypt]>=1.7.4,<2.0.0",
  "bcrypt==3.2.2",
  "python-multipart>=0.0.9,<1.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "celery>=5.4.0,<6.0.0",
  "redis>=5.0.7,<6.0.0",
  "httpx>=0.27.0,<1.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0"
]
ml = [
  "torch>=2.1.0,<3.0.0",
  "torchvision>=0.16.0,<1.0.0",
  "torchaudio>=2.1.0,<3.0.0",
  "transformers>=4.35.0,<5.0.0",
  "tokenizers>=0.15.0,<1.0.0",
  "timm>=0.9.0,<2.0.0",
  "datasets>=2.14.0,<3.0.0",
  "opencv-python-headless>=4.8.0,<5.0.0",
  "pillow>=10.0.0,<11.0.0",
  "librosa>=0.10.0,<1.0.0",
  "pydub>=0.25.0,<1.0.0",
  "soundfile>=0.12.0,<1.0.0",
  "numpy>=1.24.0,<3.0.0",
  "pandas>=2.1.0,<3.0.0",
  "scipy>=1.11.0,<2.0.0",
  "scikit-learn>=1.3.0,<2.0.0"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
